Macro Fix Again
Added Macros so that there are no cute bitwise & operations when setting the piece. It either sets to ta black_rook or a white_rook, etc. (BLACK_ROOK's value is the exact same as ROOK, I just felt that defining two macros was easier for readability in certain situations like when checking specifically for a piece when color is irrelevant)