Added Move.h, Completed Move Generation for Rooks!
So I decided to store all the availables moves in an 8x8 array (which also stores an array of available moves, so this is a 3D indexed array)

The first 2 indexes in the array correspond to the starting location of the space (ie the I and J components of the board), and for every space we can store each move set for the rooks/bishops/knights/queens etc using our movement generation function.

Because the board has to be initialized before this can work (ie the boardposition needs to be initialized for each space), I decided to create a subroutine after "setspace" in makeboard() in board.c just so that it's easier to read, but all the bulk of the work happens in 'move.c'. We can probably use this to add the bishop and knight moves as well.

I confirmed in Debugger that the values for each space are working as intended and I fixed any segmentation faults too.

For clarity, the debugger will show you the values, and the values are read as follow: RIGHT, LEFT, UP, DOWN for rook values. Meaning it will print all the right moves first, followed by the left moves, and then the up and down moves respectively in that order. I'll have to later add functionality in the moveGen function to detect collisions between other pieces but I think that will be relatively simple unless there are easier options. Enjoy!