Added Initialization for Bishop and Knight
Just checked through the debugger and the bishop and knight precompiled moveset is complete! cheers! MoveGen will take the locations stored here and figure out the i and j components corresponding to the space on the board and then check if the space is occupied. If the moveGen finds that there is -1 in any part of the move array that means that there are no more legal moves to check for that particular piece.