Finished beginning movegeneration for Knights.
Finished MoveGeneration for Knights. Currently generates all possible moves into a move list with 4 different elements: Piece type, Start Location, End location, and captured piece.

If no capture, this parameter is -1 by default.

In addition the number of moves is kept track and incremented whenever a new move is added to the list.

Final thought, I added a field within the move structure to keep track of which pieces each side has and where they are located. This is to help figure out the starting locations and generate our moves faster since our movegeneration function will only generate moves from spaces that are clearly occupied and added to the list.
Whenever a move is made I will need to Update this list, which can be done by searching for a value on the list that is the same piece type and has the same starting location as the piece that is being moved.

Will need to add support for pawns and rooks.