Mostly done for MoveGeneration, Need to fix error with the black piece Move Generation
Most of the Move Generation is done. The program can print out the board to check new board states after it makes a move. I also reset thecount to rewrite into the Move array for testing purposes. In the future we will just malloc another move List for every node and also make sure to free the movelist memory when we are done with it.

There is an error where the rook is somehow moving on the black side where it collides with a friendly black piece and somehow thinks it can occupy the space.