Minor bug fix, Movegen works for black after the first pawn movement. Need to test for other boardstates to verify bishops/rooks/queens/captures
fixed minor bug that was causing incorrect moves to be displayed. MoveGeneration needs to have support for en passant and castling, but I will try to get that done asap.

We can then use this file as a base for generating our moves within a decision tree.

Other future fixes:

1) Will need to dynamically allocate moves, because I don't really want to allocate a 100 unit array everytime I make an  array and will need to free memory after branches have been explored.

2) Castling/Pawn en Passant/Pawn Promotion needs to be called, I can use the move function using the capturedPiece variable as an exception flag to highlight special cases such as moves for castling and pawn promotion/pawn en passant.