Few Bugs in movement tree, starting point:
Tree is created with a maximum depth to traverse moves. In the move tree function I look through every possible move combination and iterate through. If the tree reaches a leaf node it runs the board evaluation function and returns the value.

Currently I need to fix bugs involving the movement generation that occur when the tree is called and moves are generated. There is currently a weird bug where undoing certain pawn moves ends up spawning pawns in both locations inadvertantly. Odds are that bug is in the unmake move function