Bug Fixes for The Tree: 2/26/2018.
So far depth up to 1 or 2 works, 3 is unclear mostly becasue I need to double check my captures, and check the # of possible captures.

In addition, there are still definitely small bugs inside the giant tree, but I ironed out a few major bugs. As far as I know it is possible that perft3 is not working becasue of interactions with the capture functions since there should be 34 unique captures. I'll need to put a counter to track that to make sure that each counter is incremented properly just in case.

You should be able to apply the board evalutation function by running it where the movetree function returns after it reaches a certain depth. From there you can keep track of your min and max values and then use alpha beta pruning to skip certain paths and ultimately return the best move. (HINT: Keep track of the current best move as well as the value and then after the function exits in the while loop in main, we can make the move using the returned move.)