Can create board from strtok'ing FEN notation
Updated my setBoard function to be able to use STRTOK to create a board state based on any given string.

Need to update to add board rights to the board from the string tokens.

In addition, need to add check legality for moves that put own king into check since they are not legal and should not be considered part of the movelist in my movegeneration.

Then I must first make sure that I add a counter for whenever a capture occurs as well as any type of check or checkmate occurs. Will need to add that to the move generation since all I know is capture.