Fixed Board Position 3: 99.9% legal up to Depth 6:
Need to implement En passant rules for Pawns, as well as promotions. But So far for Depths 1 and 2 the # of legal moves is correct according to the expected perft number for the respective depths.

For Depth 3 there are 6 illegal moves that are undetected out of ~2200 actual legal moves. I do not know if I should investigate until after I implement en passant.