Most of Promotion Laid Out, need to bug fix for depth 6
Need to bug fix for depth 6 on promotion, odds are I should use a board that promotes early since the problem is most likely tied somewhere to black pawn promotions. If not I shall look to make a board that tests these edge cases at lower depths to ensure the value is working properly.