Promotion Complete, Need to finish Castling
Need to finish Castling.

Laid out castling code for makemove, unmake move, and the necessary code within the move tree to check for the legality of the move for each of the possible cases.

Need to check movegen for possible bugs as there should be a legal castle at depth 1 and also need to update the movespaces whenever a castle occurs.

In addition, need to set the castle rights whenever a king moves and unmoves for the first time, possibly thinking about putting a flag for this to restore castling rights whenever a king moves/unmoves?