Need to finish Restoring Castling Rights
Successfully implemented the board implementation of Castling at depth 1. Need to properly restore castling rights and finish function to restore the proper castling rights based on the history. (In unmake, if the depth is <= 1, we need to set the castling rights high no matter what?) (Or possibly just implement the history anyways since we restore the rights and then break the for loop before any calculations are completed?)