Got Playable Game on board:
Board currently plays itself into a draw but it confirms to play ok at depth 5. It plays itself so it passes priority back and forth and makes moves.

One thing I definitely noticed was that the Place Tables should be different depending on whose turn it is. The tables currently in eval.c are the placetables for white's perspective and they should not be the same as black's perspective. The tables need to be copied and mirrored for black's perspective.