Lot of implementation for  Hash Table
Implemented the funcitonality for XOR'ing board positions while making and unmaking moves. Initially, Zobrist takes the current board state after setting the board and XOR's each square into the 64 bit string. After this, when make or unmake move is called the respective changed spaces are XOR'd into the zobrist (usually 2 operations) per tree traversal. Need to implement function to get index by taking the modulo of the zobrist and the size of the hash table and storing the value in the transposition table.