Changed sort function from Quicksort to MergeSort:
Also increased size of transposition table to 2^24 positions. This means that the process memory is > 1 GB right now. Increasing the transposition table size greatly increased run time and performance of program. NEED TO DEFINITELY INCORPORATE MOVE ORDERING INTO THE HARDWARE since it is taking up a large portion of the tree traversal time.